meta: update github templates
This change makes an example of how a commit message translates to
a PR which is obviously great, but from a consumer perspective you
likely want this on top of a PR instead of being appended to the bottom.

Here's where simple templating would be great -- for instance:
```html

<!-- $message -->

other stuff
```

..would output this message on top of 'other stuff' instead of being
appended to the bottom.